pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "world-model"
version = "0.1.0"
description = "Render-video-conditioned Wan 2.1 I2V for world-model-style video training (Diffusers)."
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
authors = [{ name = "World Model Training" }]
keywords = ["diffusion", "wan", "i2v", "world-model", "diffusers"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "torch>=2.1.0",
  "diffusers>=0.32.0",
  "transformers>=4.40.0",
  "accelerate>=0.30.0",
  "pandas>=2.0.0",
  "pillow>=10.0.0",
  "numpy>=1.24.0",
  "imageio>=2.31.0",
  "imageio-ffmpeg>=0.4.0",
  "safetensors>=0.4.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.4"]

[project.scripts]
world-model-train = "world_model.wan_flow.train:main"
world-model-train-fsdp = "world_model.wan_flow.train_fsdp:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
